+++ Assets/Scripts/Model/FarmItem/IdleState.cs
@@ -7,7 +7,7 @@
             _farmItem = farmItem;
         }
 
-        public override void Tick(float deltaTime)
+        public override void Tick(int deltaTime)
         {
             //
         }
